backend/requirements.txt
fastapi==0.99.1
uvicorn==0.23.2
google-generativeai==0.7.2
chromadb==0.5.3
python-multipart==0.0.9
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==1.10.26
typing-extensions>=4.8.0
google-cloud-texttospeech==2.14.1
rank_bm25==0.2.2